+++ crates/extension-plugin/src/extension_registry_actions.rs
@@ -2,8 +2,10 @@ use crate::extension_plugin::ExtensionPlugin;
 use crate::extension_registry::ExtensionRegistry;
 use moon_pdk_api::{
     ExtendProjectGraphInput, ExtendProjectGraphOutput, ExtendTaskCommandInput,
-    ExtendTaskCommandOutput, ExtendTaskScriptInput, ExtendTaskScriptOutput,
+    ExtendTaskCommandOutput, ExtendTaskScriptInput, ExtendTaskScriptOutput, SyncOutput,
+    SyncProjectInput, SyncWorkspaceInput,
 };
+use moon_plugin::CallResult;
 
 impl ExtensionRegistry {
     pub async fn extend_project_graph_all<InFn>(
@@ -18,7 +20,7 @@ impl ExtensionRegistry {
                 "extend_project_graph",
                 self.get_plugin_ids(),
                 input_factory,
-                |toolchain, input| async move { toolchain.extend_project_graph(input).await },
+                |extension, input| async move { extension.extend_project_graph(input).await },
             )
             .await?;
 
@@ -37,7 +39,7 @@ impl ExtensionRegistry {
                 "extend_task_command",
                 self.get_plugin_ids(),
                 input_factory,
-                |toolchain, input| async move { toolchain.extend_task_command(input).await },
+                |extension, input| async move { extension.extend_task_command(input).await },
             )
             .await?;
 
@@ -56,10 +58,42 @@ impl ExtensionRegistry {
                 "extend_task_script",
                 self.get_plugin_ids(),
                 input_factory,
-                |toolchain, input| async move { toolchain.extend_task_script(input).await },
+                |extension, input| async move { extension.extend_task_script(input).await },
             )
             .await?;
 
         Ok(results.into_iter().map(|result| result.output).collect())
+    }
+
+    pub async fn sync_project_all<InFn>(
+        &self,
+        input_factory: InFn,
+    ) -> miette::Result<Vec<CallResult<ExtensionPlugin, SyncOutput>>>
+    where
+        InFn: Fn(&ExtensionRegistry, &ExtensionPlugin) -> SyncProjectInput,
+    {
+        self.call_func_all(
+            "sync_project",
+            self.get_plugin_ids(),
+            input_factory,
+            |extension, input| async move { extension.sync_project(input).await },
+        )
+        .await
+    }
+
+    pub async fn sync_workspace_all<InFn>(
+        &self,
+        input_factory: InFn,
+    ) -> miette::Result<Vec<CallResult<ExtensionPlugin, SyncOutput>>>
+    where
+        InFn: Fn(&ExtensionRegistry, &ExtensionPlugin) -> SyncWorkspaceInput,
+    {
+        self.call_func_all(
+            "sync_workspace",
+            self.get_plugin_ids(),
+            input_factory,
+            |extension, input| async move { extension.sync_workspace(input).await },
+        )
+        .await
     }
 }