+++ CHANGELOG.md
@@ -11,4 +11,9 @@
 #### 🚀 Updates
 
 - **WASM API**
+  - Added `ExtendProjectGraphInput.extension_config` field.
+  - Added `ExtendTaskCommandInput.extension_config` field.
+  - Added `ExtendTaskScriptInput.extension_config` field.
   - Added `RegisterToolchainOutput.language` field.
+  - Added `SyncProjectInput.extension_config` field.
+  - Added `SyncWorkspaceInput.extension_config` field.